[EDIT] connexion and redirect to profile page
